+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/reflection/FieldProbe.java
@@ -170,12 +170,21 @@ public final class FieldProbe {
      * @return Whether or not the field can be modified reflectively.
      */
     public boolean canBeModifiedReflectively() {
-        if (field.isSynthetic()) {
+        if (field.isSynthetic() && !isKotlinDelegate()) {
             return false;
         }
         if (isFinal() && isStatic()) {
             return false;
         }
         return true;
+    }
+
+    /**
+     * Determines whether the field is a synthetic Kotlin delegate field.
+     *
+     * @return Whether or not the field is a synthetic Kotlin delegate field.
+     */
+    public boolean isKotlinDelegate() {
+        return field.isSynthetic() && field.getName().startsWith("$$delegate_");
     }
 }